- adjusted component's tests - moved _test_ directory into components - removed old tests
