--- client/_tests_/integration/sum.test.ts
@@ -1,7 +1,0 @@
-import sum from './sum';
-
-describe('sum module', () => {
-  test('adds 1 + 2 to equal 3', () => {
-    expect(sum(1, 2)).toBe(3);
-  });
-});
