--- client/_tests_/integration/sum.ts
@@ -1,3 +1,0 @@
-export default function multiply(a: number, b: number) {
-    return a + b;
-  }
