+++ client/components/_tests_/footer.test.tsx
@@ -1,5 +1,5 @@
 import renderer from "react-test-renderer";
-import Footer from "../../components/footer";
+import Footer from "../footer";
 
 describe("footer test", () =>{
     it('is redendring properly footer component ', () => {
